+++ app/adapters/registration.ts
@@ -1,6 +1,7 @@
 import OsfAdapter from './osf-adapter';
 
-export default class Registration extends OsfAdapter {}
+export default class Registration extends OsfAdapter {
+}
 
 declare module 'ember-data' {
     interface AdapterRegistry {