+++ app/locales/en/translations.ts
@@ -536,9 +536,12 @@ export default {
     institutions: {
         description: 'OSF Institutions is a free scholarly web tool that enhances transparency, fosters collaboration, and increases the visibility of research outputs at the institutional level.',
         read_more: 'Read more',
-        footer: 'Interested in setting up an OSF Institutions page for your research institution? <a href=\'mailto:contact@osf.io\'>Contact us</a>',
+        footer: 'Interested in setting up an OSF Institutions page for your research institution? <a href=\'https://cos.io/contact\'>Contact us</a>',
         title: 'Institutions',
         search_placeholder: 'Search institutions',
+    },
+    registration: {
+        project_alert: 'This registration is a frozen, non-editable version of <a class="link-solid" href="/{{projectId}}/">this project</a>',
     },
     list: {
         two_item: {