+++ app/resolve-guid/forks/route.ts
@@ -1,0 +1,3 @@
+import ResolveGuid from '../route';
+
+export default class ResolveGuidForks extends ResolveGuid {}