+++ app/router.ts
@@ -79,15 +79,17 @@ Router.map(function() {
      * will ask the API what type of object the guid refers to, then transition
      * to the appropriate `guid-<type>` route below.
      *
-     * Nested routes that begin with a guid should be unique across all types.
-     * Do not add duplicate nested routes, like `guid-node/quickfiles`.
+     * Non-unique nested routes that begin with a guid need special handling.
+     * See `resolve-guid.forks` for an example.
      */
     this.route('guid-file', { path: '/:file_guid' });
     this.route('guid-node', { path: '/:node_guid' }, function() {
         this.route('forks');
     });
     this.route('guid-preprint', { path: '/:preprint_guid' });
-    this.route('guid-registration', { path: '/:registration_guid' });
+    this.route('guid-registration', { path: '/:registration_guid' }, function() {
+        this.route('forks');
+    });
     this.route('guid-user', { path: '/:user_guid' }, function() {
         this.route('quickfiles');
     });
@@ -96,6 +98,7 @@ Router.map(function() {
     // and all the `guid-*` routes above), URLs that match will resolve to the
     // route defined last. It's very intuitive.
     this.route('resolve-guid', { path: '/:guid' });
+    this.route('resolve-guid.forks', { path: '/:guid/forks' });
 
     // Error routes
     this.route('error-no-api', { path: '*no_api_path' });